profile screen's styles added to seperate file
